--- concorrencia/v2/CheckWebsites.go
@@ -1,24 +1,0 @@
-package concurrency
-
-import (
-	"time"
-)
-
-// WebsiteChecker checks a url, returning a bool
-type WebsiteChecker func(string) bool
-
-// CheckWebsites takes a WebsiteChecker and a slice of urls and returns  a map
-// of urls to the result of checking each url with the WebsiteChecker function
-func CheckWebsites(wc WebsiteChecker, urls []string) map[string]bool {
-	results := make(map[string]bool)
-
-	for _, url := range urls {
-		go func(u string) {
-			results[u] = wc(u)
-		}(url)
-	}
-
-	time.Sleep(2 * time.Second)
-
-	return results
-}
