+++ concorrencia/v2/VerificaWebsites_test.go
@@ -1,33 +1,33 @@
-package concurrency
+package concorrencia
 
 import (
 	"reflect"
 	"testing"
 )
 
-func mockWebsiteChecker(url string) bool {
+func mockVerificadorWebsite(url string) bool {
 	if url == "waat://furhurterwe.geds" {
 		return false
 	}
 	return true
 }
 
-func TestCheckWebsites(t *testing.T) {
+func TestVerificaWebsites(t *testing.T) {
 	websites := []string{
 		"http://google.com",
 		"http://blog.gypsydave5.com",
 		"waat://furhurterwe.geds",
 	}
 
-	want := map[string]bool{
+	esperado := map[string]bool{
 		"http://google.com":          true,
 		"http://blog.gypsydave5.com": true,
 		"waat://furhurterwe.geds":    false,
 	}
 
-	got := CheckWebsites(mockWebsiteChecker, websites)
+	resultado := VerificaWebsites(mockVerificadorWebsite, websites)
 
-	if !reflect.DeepEqual(want, got) {
-		t.Fatalf("Wanted %v, got %v", want, got)
+	if !reflect.DeepEqual(esperado, resultado) {
+		t.Fatalf("esperado %v, resultado %v", esperado, resultado)
 	}
 }
